Terminate normally if the daemon is not running
The fact that dnsmasq is installed but the daemon is not running cannot
be construed as an error, therefore the logrotate script should not error
out if the daemon is not running, it should terminate normally.
The daemon not running message will now only be emitted if the verbose
option is set.

Signed-off-by: Mark Grant <m.grant.prg@gmail.com>