bash: man: Always update if new logfile location
The pid number and logfile location are ascertained on every invocation
of the script. However, a changed logfile location is only exported to
the logrotate control file if --setup is specified on the CL. This is
nonsense. If the logfile location has changed then the logrotate control
file should always be updated.

So change to always update on logfile location modification  which makes
the --setup CL option meaningless, so remove it.

Signed-off-by: Mark Grant <m.grant.prg@gmail.com>